Revert "merge with platdev"
This reverts commit a6821902974ab607de1a91f120e2f80715337e63.

Turns out merge squash will not work in this case.

+++ plugin/legacy_sound/pack.h
@@ -54,7 +54,7 @@ class legacy_pack {
 	std::string current_filename;
 	pack_open_mode open_mode;
 	std::string pack_ident;
-	uint64_t file_offset; // Offset into opened file where pack is contained, used for embedding packs into executables.
+	unsigned int file_offset; // Offset into opened file where pack is contained, used for embedding packs into executables.
 	int RefCount;
 public:
 	unsigned int next_stream_idx;
